feat(delete dialog) : added dialog for delete
also
  modify date range to 15 / 2 months
  fix imei tutorial page from home

feat #12

+++ lib/absen_manual/absen_manual_list/presentation/absen_manual_dtl_dialog.dart
@@ -279,19 +279,27 @@ class AbsenManualDtlDialog extends ConsumerWidget {
                     TappableSvg(
                         assetPath: Assets.iconDelete,
                         onTap: () {
-                          ref
-                              .read(createAbsenManualNotifierProvider.notifier)
-                              .deleteAbsenmanual(
-                                idAbsenmnl: item.idAbsenmnl!,
-                                onError: (msg) {
-                                  return DialogHelper.showCustomDialog(
-                                    msg,
-                                    context,
-                                  ).then((_) => ref
-                                      .read(errLogControllerProvider.notifier)
-                                      .sendLog(errMessage: msg));
-                                },
-                              );
+                          context.pop();
+                          return DialogHelper.showConfirmationDialog(
+                              context: context,
+                              label: 'Hapus form ? ',
+                              onPressed: () async {
+                                context.pop();
+                                await ref
+                                    .read(createAbsenManualNotifierProvider
+                                        .notifier)
+                                    .deleteAbsenmanual(
+                                        idAbsenmnl: item.idAbsenmnl!,
+                                        onError: (msg) {
+                                          return DialogHelper.showCustomDialog(
+                                            msg,
+                                            context,
+                                          ).then((_) => ref
+                                              .read(errLogControllerProvider
+                                                  .notifier)
+                                              .sendLog(errMessage: msg));
+                                        });
+                              });
                         })
                 ],
               )
