+++ lib/cuti/cuti_list/presentation/cuti_dtl_dialog.dart
@@ -332,18 +332,25 @@ class CutiDtlDialog extends ConsumerWidget {
                     assetPath: Assets.iconDelete,
                     onTap: () async {
                       context.pop();
-                      await ref
-                          .read(createCutiNotifierProvider.notifier)
-                          .deleteCuti(
-                              idCuti: item.idCuti!,
-                              onError: (msg) {
-                                return DialogHelper.showCustomDialog(
-                                  msg,
-                                  context,
-                                ).then((_) => ref
-                                    .read(errLogControllerProvider.notifier)
-                                    .sendLog(errMessage: msg));
-                              });
+                      return DialogHelper.showConfirmationDialog(
+                          context: context,
+                          label: 'Hapus form ? ',
+                          onPressed: () async {
+                            context.pop();
+                            await ref
+                                .read(createCutiNotifierProvider.notifier)
+                                .deleteCuti(
+                                    idCuti: item.idCuti!,
+                                    onError: (msg) {
+                                      return DialogHelper.showCustomDialog(
+                                        msg,
+                                        context,
+                                      ).then((_) => ref
+                                          .read(
+                                              errLogControllerProvider.notifier)
+                                          .sendLog(errMessage: msg));
+                                    });
+                          });
                     },
                   )
               ],
