+++ lib/ganti_hari/ganti_hari_list/presentation/ganti_hari_dtl_dialog.dart
@@ -324,18 +324,26 @@ class GantiHariDtlDialog extends ConsumerWidget {
                     TappableSvg(
                         assetPath: Assets.iconDelete,
                         onTap: () {
-                          ref
-                              .read(createGantiHariProvider.notifier)
-                              .deleteGantiHari(
-                                  idDayOff: item.idDayOff!,
-                                  onError: (msg) {
-                                    return DialogHelper.showCustomDialog(
-                                      msg,
-                                      context,
-                                    ).then((_) => ref
-                                        .read(errLogControllerProvider.notifier)
-                                        .sendLog(errMessage: msg));
-                                  });
+                          context.pop();
+                          return DialogHelper.showConfirmationDialog(
+                              context: context,
+                              label: 'Hapus form ? ',
+                              onPressed: () async {
+                                context.pop();
+                                await ref
+                                    .read(createGantiHariProvider.notifier)
+                                    .deleteGantiHari(
+                                        idDayOff: item.idDayOff!,
+                                        onError: (msg) {
+                                          return DialogHelper.showCustomDialog(
+                                            msg,
+                                            context,
+                                          ).then((_) => ref
+                                              .read(errLogControllerProvider
+                                                  .notifier)
+                                              .sendLog(errMessage: msg));
+                                        });
+                              });
                         })
                 ],
               )
