+++ lib/izin/izin_list/presentation/izin_dtl_dialog.dart
@@ -326,18 +326,25 @@ class IzinDtlDialog extends ConsumerWidget {
                       assetPath: Assets.iconDelete,
                       onTap: () async {
                         context.pop();
-                        await ref
-                            .read(createIzinNotifierProvider.notifier)
-                            .deleteIzin(
-                                idIzin: item.idIzin!,
-                                onError: (msg) {
-                                  return DialogHelper.showCustomDialog(
-                                    msg,
-                                    context,
-                                  ).then((_) => ref
-                                      .read(errLogControllerProvider.notifier)
-                                      .sendLog(errMessage: msg));
-                                });
+                        return DialogHelper.showConfirmationDialog(
+                            context: context,
+                            label: 'Hapus form ? ',
+                            onPressed: () async {
+                              context.pop();
+                              await ref
+                                  .read(createIzinNotifierProvider.notifier)
+                                  .deleteIzin(
+                                      idIzin: item.idIzin!,
+                                      onError: (msg) {
+                                        return DialogHelper.showCustomDialog(
+                                          msg,
+                                          context,
+                                        ).then((_) => ref
+                                            .read(errLogControllerProvider
+                                                .notifier)
+                                            .sendLog(errMessage: msg));
+                                      });
+                            });
                       },
                     )
                 ],
